Resolved conflicts:
- src/Storages/MergeTree/MergeTreeIOSettings.h: Added both `use_hybrid_row_reading` and `read_only_column_sample` fields
- src/Storages/MergeTree/MergeTreeSettings.cpp: Added both hybrid storage settings and distributed_index_analysis settings

+++ base/base/wide_integer_impl.h
@@ -1406,7 +1406,7 @@ constexpr integer<Bits, Signed>::operator long double() const noexcept
         long double t = res;
         res *= static_cast<long double>(std::numeric_limits<base_type>::max());
         res += t;
-        res += tmp.items[_impl::big(i)];
+        res += static_cast<long double>(tmp.items[_impl::big(i)]);
     }
 
     if (_impl::is_negative(*this))
